changes in quizTemplate - restructed code into packages for activities and objects...reformatted code..made getters and setters in object classes and made instance variables private..
