+++ QuizTemplate/source/QuizTemplateApp/app/src/main/res/layout/questions_view.xml
@@ -2,7 +2,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="fill_parent"
     android:layout_height="wrap_content"
-    android:background="@color/BlueBg" >
+    android:background="@color/BlueBg">
 
     <TextView
         android:id="@+id/question_no"
@@ -11,8 +11,7 @@
         android:layout_alignParentTop="true"
         android:layout_margin="10dp"
         android:textAppearance="?android:attr/textAppearanceMedium"
-        android:textColor="@android:color/white" >
-    </TextView>
+        android:textColor="@android:color/white"></TextView>
 
     <View
         android:id="@+id/divider"
@@ -21,8 +20,7 @@
         android:layout_below="@id/question_no"
         android:layout_marginLeft="10dp"
         android:layout_marginRight="10dp"
-        android:background="#8798AC" >
-    </View>
+        android:background="#8798AC"></View>
 
     <TextView
         android:id="@+id/question_label"
@@ -31,15 +29,14 @@
         android:layout_below="@id/divider"
         android:layout_margin="10dp"
         android:textAppearance="?android:attr/textAppearanceMedium"
-        android:textColor="@android:color/white" >
-    </TextView>
+        android:textColor="@android:color/white"></TextView>
 
     <RadioGroup
         android:id="@+id/radioGroup1"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
         android:layout_below="@id/question_label"
-        android:layout_margin="10dp" >
+        android:layout_margin="10dp">
 
         <RadioButton
             android:id="@+id/radio0"
@@ -47,8 +44,7 @@
             android:layout_height="wrap_content"
             android:checked="false"
             android:onClick="radioClick"
-            android:textColor="@android:color/white" >
-        </RadioButton>
+            android:textColor="@android:color/white"></RadioButton>
 
         <RadioButton
             android:id="@+id/radio1"
@@ -56,8 +52,7 @@
             android:layout_height="wrap_content"
             android:checked="false"
             android:onClick="radioClick"
-            android:textColor="@android:color/white" >
-        </RadioButton>
+            android:textColor="@android:color/white"></RadioButton>
 
         <RadioButton
             android:id="@+id/radio2"
@@ -66,8 +61,7 @@
             android:checked="false"
             android:onClick="radioClick"
             android:text="RadioButton"
-            android:textColor="@android:color/white" >
-        </RadioButton>
+            android:textColor="@android:color/white"></RadioButton>
 
         <RadioButton
             android:id="@+id/radio3"
@@ -75,15 +69,14 @@
             android:layout_height="wrap_content"
             android:checked="false"
             android:onClick="radioClick"
-            android:textColor="@android:color/white" >
-        </RadioButton>
+            android:textColor="@android:color/white"></RadioButton>
     </RadioGroup>
 
     <LinearLayout
         android:id="@+id/ll"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true" >
+        android:layout_alignParentBottom="true">
 
         <Button
             android:id="@+id/next_button"
@@ -93,8 +86,7 @@
             android:layout_weight="1"
             android:padding="10dp"
             android:text="Next"
-            android:textColor="@android:color/white" >
-        </Button>
+            android:textColor="@android:color/white"></Button>
 
         <Button
             android:id="@+id/submit_button"
@@ -104,8 +96,7 @@
             android:layout_weight="1"
             android:padding="10dp"
             android:text="Submit"
-            android:textColor="@android:color/white" >
-        </Button>
+            android:textColor="@android:color/white"></Button>
     </LinearLayout>
 
 </RelativeLayout>
