+++ QuizTemplate/source/QuizTemplateApp/app/src/main/res/layout/start_view.xml
@@ -6,8 +6,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity" >
+    android:paddingTop="@dimen/activity_vertical_margin">
 
     <View
         android:id="@+id/divider"
@@ -26,8 +25,7 @@
         android:gravity="center"
         android:textAppearance="?android:attr/textAppearanceLarge"
         android:textColor="@android:color/white"
-        android:textStyle="bold" >
-    </TextView>
+        android:textStyle="bold"></TextView>
 
     <TextView
         android:id="@+id/tv_author"
@@ -38,8 +36,7 @@
         android:layout_margin="10dp"
         android:gravity="center"
         android:textAppearance="?android:attr/textAppearanceLarge"
-        android:textColor="@android:color/white" >
-    </TextView>
+        android:textColor="@android:color/white"></TextView>
 
     <Button
         android:id="@+id/btn_start"
@@ -47,11 +44,10 @@
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
-        android:text="@string/start"
         android:paddingLeft="15dp"
         android:paddingRight="15dp"
+        android:text="@string/start"
         android:textAppearance="@android:style/TextAppearance.Large"
-        android:textColor="@android:color/white" >
-    </Button>
+        android:textColor="@android:color/white"></Button>
 
 </RelativeLayout>
